+++ jest.config.js
@@ -2,11 +2,14 @@ module.exports = {
   moduleNameMapper: {
     "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$":
       "<rootDir>/__mocks__/fileMock.js",
-    "\\.(css|less|scss)$": "<rootDir>/__mocks__/styleMock.js"
+    "\\.(css|less|scss)$": "<rootDir>/__mocks__/styleMock.js",
+    "^@components(.*)$": "<rootDir>/src/components$1",
+    "^@constants(.*)$": "<rootDir>/src/constants$1",
+    "^@assets(.*)$": "<rootDir>/src/assets$1",
+    "^@hocs(.*)$": "<rootDir>/src/hocs$1"
   },
   moduleFileExtensions: ["js", "jsx", "json"],
   transform: {
     "^.+\\.(js|jsx)?$": "babel-jest"
-  },
-  transformIgnorePatterns: ["<rootDir>/node_modules/"]
+  }
 };
