--- src/components/avatar/Avatar.scss
@@ -1,8 +1,0 @@
-.avatar__img {
-  width: 300px;
-  height: 300px;
-  object-fit: cover;
-  border-radius: 50%;
-  display: block;
-  margin: 32px auto;
-}
