+++ src/components/common/badge/Badge.scss
