+++ src/components/common/index.js
@@ -1,0 +1,4 @@
+import Badge from "./badge/Badge";
+import Skill from "./skill/Skill";
+
+export { Badge, Skill };
