+++ src/components/common/skill/Skill.scss
@@ -1,4 +1,4 @@
-@import "../../assets/styles/grid.scss";
+@import "@assets/styles/grid.scss";
 
 .skill__article {
   @include content-grid;
