+++ src/components/sections/index.js
@@ -1,0 +1,4 @@
+import SkillsSection from "./skillsSection/SkillsSection";
+import SplashSection from "./splashSection/SplashSection";
+
+export { SkillsSection, SplashSection };
