+++ src/components/sections/skillsSection/SkillsSection.scss
@@ -1,0 +1,5 @@
+@import "@assets/styles/section.scss";
+
+.skills__section {
+  @include section;
+}
