+++ src/components/sections/splashSection/SplashSection.scss
@@ -1,5 +1,5 @@
-@import "../../../assets/styles/grid.scss";
-@import "../../../assets/styles/section.scss";
+@import "@assets/styles/grid.scss";
+@import "@assets/styles/section.scss";
 
 .splash-section__section {
   @include section;
