--- src/components/skills/Skills.scss
@@ -1,5 +1,0 @@
-@import "../../assets/styles/section.scss";
-
-.skills__section {
-  @include section;
-}
