+++ src/components/skills/backendSkill/BackendSkill.scss
@@ -1,4 +1,4 @@
-@import "../../../assets/styles/badge.scss";
+@import "@assets/styles/badge.scss";
 
 .backend-skill__article {
   .badges__div {
