+++ src/components/skills/index.js
@@ -1,0 +1,6 @@
+import BackendSkill from "./backendSkill/BackendSkill";
+import FrontendSkill from "./frontendSkill/FrontendSkill";
+import InfrastructureSkill from "./infrastructureSkill/InfrastructureSkill";
+import UxSkill from "./uxSkill/UxSkill";
+
+export { BackendSkill, FrontendSkill, InfrastructureSkill, UxSkill };
